+++ changes.md
@@ -1,3 +1,5 @@
-
 ## DegreeSign Server Cache SDK - Change Log
 
+### 14 December 2025 (version 0.0.7)
+* Added `wrtJCompressed` and `redJCompressed` for saving JSON in compressed format
+* Error logs polished